refactor[#40] : entities/series 에 inde.ts 에 존재하던 모든 api 코드 하나의 파일로 모듈화
