+++ blog/src/entities/series/model/seriesQueryKey.ts
@@ -1,0 +1,3 @@
+export const SERIES_QUERY_KEY = {
+  default: () => ["series"] as const
+} as const;
